+++ src/test/java/io/tarantool/driver/integration/RetryingTarantoolTupleClientIT.java
@@ -1,26 +1,38 @@
 package io.tarantool.driver.integration;
 
+import io.tarantool.driver.api.TarantoolClient;
 import io.tarantool.driver.api.TarantoolClientConfig;
+import io.tarantool.driver.api.TarantoolClientFactory;
+import io.tarantool.driver.api.TarantoolResult;
 import io.tarantool.driver.api.retry.TarantoolRequestRetryPolicies;
+import io.tarantool.driver.api.tuple.TarantoolTuple;
 import io.tarantool.driver.auth.SimpleTarantoolCredentials;
 import io.tarantool.driver.core.ClusterTarantoolTupleClient;
 import io.tarantool.driver.core.ProxyTarantoolTupleClient;
 import io.tarantool.driver.core.RetryingTarantoolTupleClient;
 import io.tarantool.driver.exceptions.TarantoolAttemptsLimitException;
+import io.tarantool.driver.exceptions.TarantoolClientException;
 import io.tarantool.driver.exceptions.TarantoolFunctionCallException;
 import io.tarantool.driver.exceptions.TarantoolInternalException;
 import io.tarantool.driver.exceptions.TarantoolTimeoutException;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
+import java.sql.Timestamp;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionException;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.Predicate;
 
+import static java.lang.Math.abs;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -216,6 +228,102 @@ public class RetryingTarantoolTupleClientIT extends SharedCartridgeContainer {
             assertEquals(TarantoolTimeoutException.class, e.getCause().getClass());
             assertTrue(e.getCause().getMessage().contains("Operation timeout value exceeded after"));
             assertEquals(TimeoutException.class, e.getCause().getCause().getClass());
+        }
+    }
+
+    void longTermFunction(TarantoolClient<TarantoolTuple, TarantoolResult<TarantoolTuple>> client) {
+        ArrayList<CompletableFuture<List<?>>> list = new ArrayList<>(1000);
+        int batchSize = 1000;
+
+        for (int i = 0; i < 100; i++) {
+            for (int j = 0; j < batchSize; j++) {
+                CompletableFuture<List<?>> future = client
+                    .call("simple_long_running_function", Collections.singletonList(0.01));
+                list.add(j, future);
+            }
+            for (int j = 0; j < batchSize; j++) {
+                assertEquals(list.get(j).join().get(0), true);
+            }
+            list.clear();
+        }
+    }
+
+    @Test
+    void testAsyncRetryingPerformance() throws InterruptedException, ExecutionException, TimeoutException {
+        TarantoolClient<TarantoolTuple, TarantoolResult<TarantoolTuple>> asyncClient
+            = TarantoolClientFactory.createClient()
+            .withCredentials(new SimpleTarantoolCredentials(USER_NAME, PASSWORD))
+            .withAddress(container.getRouterHost(), container.getRouterPort())
+            .withConnectTimeout(1000)
+            .withReadTimeout(1000)
+            .withRetryingIndefinitely(policy -> policy.withDelay(500)
+                .withRequestTimeout(2000)
+                .withOperationTimeout(2000))
+            .build();
+
+        TarantoolClient<TarantoolTuple, TarantoolResult<TarantoolTuple>> client
+            = TarantoolClientFactory.createClient()
+            .withCredentials(new SimpleTarantoolCredentials(USER_NAME, PASSWORD))
+            .withAddress(container.getRouterHost(), container.getRouterPort())
+            .withConnectTimeout(1000)
+            .withReadTimeout(1000)
+            .build();
+
+        CompletableFuture<Boolean> syncFuture = new CompletableFuture<>();
+        CompletableFuture<Boolean> asyncFuture = new CompletableFuture<>();
+
+        AtomicReference<Timestamp> asyncTimestamp = new AtomicReference<>();
+        AtomicReference<Timestamp> syncTimestamp = new AtomicReference<>();
+
+        CompletableFuture<Void> asyncFutureTimestamp =
+            asyncFuture.thenRun(() -> asyncTimestamp.set(new Timestamp(System.currentTimeMillis())));
+        CompletableFuture<Void> syncFutureTimestamp =
+            syncFuture.thenRun(() -> syncTimestamp.set(new Timestamp(System.currentTimeMillis())));
+
+        new Thread(() -> {
+            longTermFunction(client);
+            syncFuture.complete(true);
+        }).start();
+        new Thread(() -> {
+            longTermFunction(asyncClient);
+            asyncFuture.complete(true);
+        }).start();
+
+        CompletableFuture.allOf(asyncFutureTimestamp, syncFutureTimestamp).get(30, TimeUnit.SECONDS);
+        assertTrue(abs(asyncTimestamp.get().getTime() - asyncTimestamp.get().getTime()) < 500);
+    }
+
+    @Test
+    public void test_incorrectCallback_shouldHandleCorrectly() {
+        //given
+        Predicate<Throwable> callbackWithException = e -> {
+            // it's possible, when user will specify incorrect callback
+            // such as throwable.getMessage().equals() and message is null
+            throw new NullPointerException();
+        };
+
+        //when
+        List<TarantoolClient<TarantoolTuple, TarantoolResult<TarantoolTuple>>> clients = Arrays.asList(
+            TarantoolClientFactory.createClient()
+                .withAddress(container.getRouterHost(), container.getRouterPort())
+                .withCredentials(USER_NAME, PASSWORD)
+                .withRetryingByNumberOfAttempts(5, callbackWithException, policy -> policy)
+                .build(),
+            TarantoolClientFactory.createClient()
+                .withAddress(container.getRouterHost(), container.getRouterPort())
+                .withCredentials(USER_NAME, PASSWORD)
+                .withRetryingIndefinitely(callbackWithException, policy -> policy)
+                .build()
+        );
+
+        for (TarantoolClient client :
+            clients) {
+            //then
+            CompletionException completionException =
+                assertThrows(CompletionException.class, () -> client.eval("return error").join());
+            assertTrue(completionException.getCause() instanceof TarantoolClientException);
+            assertTrue(completionException.getMessage()
+                .contains("Specified in TarantoolClient predicate for exception check threw exception: "));
         }
     }
 }